requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.72.0
h11==0.12.0
httptools==0.3.0
idna==3.3
MarkupSafe==2.0.1
Pillow==9.0.0
pydantic==1.9.0
pymongo==4.0.1
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
